+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rrze-univis",
-  "version": "3.7.8-3",
+  "version": "3.7.8-4",
   "description": "Einbindung von Daten aus UnivIS",
   "main": "rrze-univis.php",
   "textdomain": "rrze-univis",
