+++ readme.txt
@@ -1,5 +1,5 @@
 === Plugin Name: rrze-univis ===
-Version: 3.7.8-3
+Version: 3.7.8-4
 Plugin URI: https://github.com/RRZE-Webteam/rrze-univis
 GitHub Issue URL: https://github.com/RRZE-Webteam/rrze-univis/issues
 Author: RRZE-Webteam <webmaster@fau.de>
